requirements.txt
streamlit>=1.54.0
pandas>=2.3.0
sqlite3
python-dotenv>=1.0.1
plotly>=5.18.0
requests>=2.31.0
pyinstaller>=6.3.0
numpy-financial>=1.0.0
